fix(formMapping): Move class
Use a different package for the FormMappingMarshaller to avoid package
collision in osgi environmemt
