+++ business-archive/src/main/java/org/bonitasoft/engine/bpm/bar/FormMappingContribution.java
@@ -20,6 +20,7 @@ import java.nio.file.Files;
 import javax.xml.bind.JAXBException;
 
 import org.bonitasoft.engine.bpm.bar.form.model.FormMappingModel;
+import org.bonitasoft.engine.bpm.form.FormMappingModelMarshaller;
 import org.xml.sax.SAXException;
 
 /**
